requirements.txt
sqlalchemy==2.0.19
pandas==2.0.3
jupyter==1.0.0
notebook==6.5.4
sqlite3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0 